move 'traverse' tests to separate file
