+++ types/ramda/ramda-tests.ts
@@ -146,15 +146,6 @@ R.useWith(addAll, [double, square]);
 };
 
 () => {
-    const of = Array.of;
-    const fn = (x: number) => Array.of(x + 1);
-    const list = [1, 2, 3];
-    R.traverse(of, fn, list);
-    R.traverse(of, fn)(list);
-    R.traverse<number, number>(of)(fn, list);
-};
-
-() => {
     const a: boolean  = R.tryCatch<boolean>(R.prop("x"), R.F)({x: true}); // => true
     const a1: boolean = R.tryCatch(R.prop<"x", true>("x"), R.F)({x: true}); // => true
     const b: boolean = R.tryCatch<boolean>(R.prop("x"), R.F)(null);      // => false
